(1) Fixed problem with incorrect PMU counter output (it turns out that enabling/disabling only the group leader, but resetting *all* others is necessary). (2) Added number of cache line writes to tracked output (as well as the percentage thereof relating to cache line owner conflicts)

+++ src/libmallocprof.h
@@ -243,8 +243,8 @@ void sampleMemoryOverhead(int, siginfo_t*, void*);
 void updateGlobalFriendlinessData();
 void calcAppFriendliness();
 
-inline ulong safeDivisor(ulong divisor) {
-	return (!divisor) ? 1 : divisor;
+inline double safeDivisor(ulong divisor) {
+	return (!divisor) ? 1.0 : (double)divisor;
 }
 
 #include "shadowmemory.hh"
